Update Strings.it.xlf by removing a spurious character
Removed a spurious Ł character. The message MSBuildVersionMessage is shown every time a compilation is done in VS Code, so it is quite annoying (and considering the strangeness of the character, not used in the Italian language, I'm quite sure many Italian programmers thought it was something "special" that was "lost" because the font they used didn't support it). Ah. I'm Italian and my first language is Italian.